Simplified mint method
And changed the mint callback. The original calculations didn't work perfectly, because the Uniswap pool uses slightly different (but still very elaborate) math for counting liquidity from amounts. Instead of recalculating it all on our end before sending, I chose to allow the owed tokens calculation to take place on their end, return to us, and then make us pay out tokens owed. This is different from the Gelato way of doing things, but seems to be more in line with what the Uniswap team had in mind when implementing callbacks.